+++ BookSaleProject/src/main/java/com/example/BookSaleProject/Model/Service/IRateService.java
@@ -3,5 +3,5 @@ package com.example.BookSaleProject.Model.Service;
 import com.example.BookSaleProject.Model.Entity.Book;
 
 public interface IRateService {
-    public double getScoreByIdBook(Book book);
+    public float getScoreByIdBook(Book book);
 }
